MAINT add base class for voting and stacking (#15084)
